Expose the Sequence object
Access to the underlying Sequence object is blocked since the
sequence function (`$.fn.sequence`) returns the iterator from
`each` and not the object itself. This commit adds an optional,
backwards-compatible parameter to the mentioned function (namely
`returnSequenceObject`), that, when passed as true, returns the
sequence object instead of the array of DOM elements. This enables
one to access functions such as startAutoPlay and stopAutoPlay.